Add files via upload
Project Description:
This project utilizes data from the 2014 Major League Baseball seasons to develop an algorithm that predicts the number of wins for a given team in the 2015 season. The prediction is based on various performance indicators, including the number of wins credited to a pitcher, runs scored, at-bats, hits, home runs, and other key statistics that impact a team's success.

The dataset consists of 16 different features, each serving as input to the machine learning model, and the output is the number of predicted wins. These features represent essential aspects of a team's performance, such as pitcher wins, runs scored, stolen bases, earned run average, and more.

To accomplish this task, we employ regression models, including Gradient Boosting Regressor and Support Vector Regressor (SVR), to capture the complex relationships between the input features and the target variable. We perform thorough data preprocessing, handle outliers and skewness, and conduct feature engineering to ensure the data is ready for modeling.

After building the models, we evaluate their performance using metrics like Mean Squared Error (MSE), Root Mean Squared Error (RMSE), and R-squared (R2) on a testing dataset. The best-performing model is selected based on its ability to generalize well and produce accurate predictions.

To access the dataset and understand the meaning of the columns representing baseball statistics, a link to the dataset and a reference to the respective baseball statistics are provided. This project aims to provide valuable insights into baseball performance and help in making informed decisions for the upcoming season.

GitHub Dataset Link:
•	https://github.com/dsrscientist/Data-Science-ML-Capstone-Projects/blob/master/baseball.csv

By leveraging machine learning techniques and baseball statistics, this project aims to enhance predictive capabilities and contribute to the exciting world of baseball analytics.